Added CommandHandling and added prefix

+++ DiscordBotCore/Storage/Implementations/JsonStorage.cs
@@ -7,16 +7,30 @@ namespace DiscordBotCore.Storage.Implementations
     {
         public T RestoreObject<T>(string key)
         {
-            var json = File.ReadAllText($"{key}.json");
-            return JsonConvert.DeserializeObject<T>(json);
+            var file = ($"{key}.json");
+            var json = File.ReadAllText(file);
+            var botcnf = JsonConvert.DeserializeObject<T>(json);
+            return botcnf;
         }
 
         public void StoreObject(object obj, string key)
         {
             var file = $"{key}.json";
-            Directory.CreateDirectory(Path.GetDirectoryName(file));
-            var json = JsonConvert.SerializeObject(obj);
-            File.WriteAllText(file, json);
+            if (!Directory.Exists(Path.GetDirectoryName(file)))
+            {
+                Directory.CreateDirectory(Path.GetDirectoryName(file));
+            }
+            if (!File.Exists(Path.GetDirectoryName(file)))
+            {
+                string json = JsonConvert.SerializeObject(obj, Formatting.Indented);
+                File.WriteAllText(file, json);
+            }
         }
+    }
+
+    public struct BotConfig
+    {
+        public string BotToken;
+        public string cmdPrefix;
     }
 }
